Strip down ReactorRepositoryManager
Currently ReactorRepositoryManager is also involved in computing target
platforms that is actually called on other places with similar context
and only referenced on single places.

This strip down ReactorRepositoryManager to providing the
PublishingRepository and inline the computation of platforms to allow
more local optimiziations.
