+++ tycho-api/src/main/java/org/eclipse/tycho/TargetPlatform.java
@@ -33,6 +33,7 @@ public interface TargetPlatform {
      * Key under which the final target platform is stored in the reactor project instances.
      */
     String FINAL_TARGET_PLATFORM_KEY = "org.eclipse.tycho.core.TychoConstants/targetPlatform";
+    String PRELIMINARY_TARGET_PLATFORM_KEY = "org.eclipse.tycho.core.TychoConstants/dependencyOnlyTargetPlatform";
 
     /**
      * Returns an artifact of the given type, id and matching version. The version reference string
@@ -77,12 +78,11 @@ public interface TargetPlatform {
 
     default ResolvedArtifactKey resolvePackage(String packageName, String versionRef)
             throws DependencyResolutionException, IllegalArtifactReferenceException {
-        ArtifactKey packageJar = resolveArtifact(PublisherHelper.CAPABILITY_NS_JAVA_PACKAGE, packageName,
-                versionRef);
-        File location = getArtifactLocation(new DefaultArtifactKey(ArtifactType.TYPE_ECLIPSE_PLUGIN,
-                packageJar.getId(), packageJar.getVersion()));
-        return ResolvedArtifactKey.of(ArtifactType.TYPE_ECLIPSE_PLUGIN, packageJar.getId(),
-                packageJar.getVersion(), location);
+        ArtifactKey packageJar = resolveArtifact(PublisherHelper.CAPABILITY_NS_JAVA_PACKAGE, packageName, versionRef);
+        File location = getArtifactLocation(
+                new DefaultArtifactKey(ArtifactType.TYPE_ECLIPSE_PLUGIN, packageJar.getId(), packageJar.getVersion()));
+        return ResolvedArtifactKey.of(ArtifactType.TYPE_ECLIPSE_PLUGIN, packageJar.getId(), packageJar.getVersion(),
+                location);
     }
 
     /**
