+++ tycho-core/src/main/java/org/eclipse/tycho/repository/registry/facade/ReactorRepositoryManager.java
@@ -13,15 +13,9 @@
  *******************************************************************************/
 package org.eclipse.tycho.repository.registry.facade;
 
-import java.util.List;
-
-import org.eclipse.tycho.ExecutionEnvironmentConfiguration;
 import org.eclipse.tycho.ReactorProject;
 import org.eclipse.tycho.ReactorProjectIdentities;
-import org.eclipse.tycho.TargetPlatform;
 import org.eclipse.tycho.p2.repository.PublishingRepository;
-import org.eclipse.tycho.p2.target.facade.PomDependencyCollector;
-import org.eclipse.tycho.p2.target.facade.TargetPlatformConfigurationStub;
 
 /**
  * Manages the p2 repositories for the projects' build results ("publishing repository") and the p2
@@ -30,41 +24,16 @@ import org.eclipse.tycho.p2.target.facade.TargetPlatformConfigurationStub;
 public interface ReactorRepositoryManager {
 
     /**
-     * Computes the target platform with dependency-only p2 metadata and attaches it to the given
-     * project.
-     * 
-     * @param project
-     *            the reactor project to compute the target platform for.
-     * @param resolvedEnvironment
-     */
-    TargetPlatform computePreliminaryTargetPlatform(ReactorProject project,
-            TargetPlatformConfigurationStub tpConfiguration, ExecutionEnvironmentConfiguration eeConfiguration,
-            List<ReactorProject> reactorProjects);
-
-    /**
-     * Computes the (immutable) target platform with final p2 metadata and attaches it to the given
-     * project.
-     * 
-     * @param project
-     *            the reactor project to compute the target platform for.
-     * @param upstreamProjects
-     *            Other projects in the reactor which have already been built and may be referenced
-     *            by the given project.
-     */
-    TargetPlatform computeFinalTargetPlatform(ReactorProject project,
-            List<? extends ReactorProjectIdentities> upstreamProjects, PomDependencyCollector pomDependencyCollector);
-
-    /**
-     * Returns the target platform with final p2 metadata for the given project.
-     */
-    TargetPlatform getFinalTargetPlatform(ReactorProject project);
-
-    /**
      * Returns the project's publishing repository.
      * 
      * @param project
      *            a reference to a project in the reactor.
+     * @return the {@link PublishingRepository} for the {@link ReactorProjectIdentities}
      */
     PublishingRepository getPublishingRepository(ReactorProjectIdentities project);
+
+    default PublishingRepository getPublishingRepository(ReactorProject project) {
+        return getPublishingRepository(project.getIdentities());
+    }
 
 }
